% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brfss}
\alias{brfss}
\title{Behavioral Risk Factor Surveillance System}
\format{A data frame
\describe{
   \item{state}{The respondent's state or entity of residence.}
   \item{year}{The interview year.}
   \item{genhealth}{General health status.}
   \item{physhealth}{Number of days in the last month that respondent's physical health was not good.}
   \item{menthealth}{Number of days in the last month that respondent's mental health was not good.}
   \item{heartattack}{Whether the respondent was ever told that they had a heart attack.}
   \item{heartdisease}{Whether the respondent was ever told that they had heart disease.}
   \item{stroke}{Whether the respondent was ever told that they had a stroke.}
   \item{asthma}{Whether the respondent was ever told that they had asthma.}
   \item{diabetes}{Whether the respondent was ever told that they had diabetes.}
   \item{sex}{The sex of the respondent.}
   \item{education}{The highest grade or year of school the respondent completed.}
   \item{veteran}{Whether the respondent has ever served on active duty in the US Armed Forces.}
   \item{income}{The respondent's annual household income level.}
   \item{weight}{The respondent's weight, in pounds.}
   \item{height}{The respondent's height, in inches.}
   \item{smokeany}{Whether the respondent has smoked at least 100 cigarettes in their entire life.}
   \item{smokeperday}{The frequency of the respondent's current smoking habit.}
   \item{alcohol}{The number of days in the past 30 days that the respondent had at least one alcoholic beverage. Some respondents provided number of days in the last week, and those values have been multiplied by 4.}
   \item{maxdrinks}{The largest number of drinks on any occasion that the respondent has consumed over the last 30 days.}
   \item{exerciseany}{Whether the respondent participated in any physical activities other than work in the last month.}
   \item{hivtest}{Whether the respondent has ever been tested for HIV.}
   \item{age}{The age of the respondent in 5-year intervals.}
   \item{age7}{The age of the respondent in 10-year intervals.}
}}
\source{
\url{https://www.cdc.gov/brfss/annual_data/annual_data.htm}
}
\usage{
brfss
}
\description{
Data from the BRFSS survey conducted annually by the \emph{Centers for Disease Control
and Prevention} (CDC). It contains a random sample of 100000 out of a total
1.3 million cases across four years (2000, 2005, 2010, 2015).
}
\details{
Data Source: \url{https://www.cdc.gov/brfss/annual_data/annual_data.htm}
}
\seealso{
\code{\link{brfss2}} for the full dataset
}
\keyword{datasets}
