% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{US County Data}
\format{
A data frame
\describe{
   \item{state}{The name of the state containing the county.}
   \item{name}{The name of the county.}
   \item{pop2020}{The 2020 county population.}
   \item{pop2010}{The 2010 county population.}
   \item{pop2000}{The 2000 county population.}
   \item{age_under_5}{Percent of population under 5 (2020).}
   \item{age_under_20}{Percent of population under 1820 (2020).}
   \item{age_over_65}{Percent of population over 65 (2020).}
   \item{female}{Percent of population that is female (2020).}
   \item{white}{Percent of population that is white (2020).}
   \item{black}{Percent of population that is black (2020).}
   \item{native}{Percent of population that is american indian or alaska native (2020).}
   \item{asian}{Percent of population that is asian (2020).}
   \item{hispanic}{Percent of population that is hispanic (2020).}
   \item{white_not_hispanic}{Percent of population that is white and not Hispanic (2020).}
   \item{foreign_born}{Percent of population that is foreign-born (2016-2020).}
   \item{foreign_spoken_at_home}{Percent of population that speaks a foreign language at home (2016-2020).}
   \item{hs_grad}{Percent of population that is a high school graduate (2016-2020).}
   \item{bachelor_degree}{Percent of population that earned a bachelor's degree (2016-2020).}
   \item{veterans}{Number of veterans in the county (2020).}
   \item{home_ownership}{Homeownership rate (2016-2020).}
   \item{households}{Number of households in the county (2020).}
   \item{persons_per_household}{Mean number of persons per household (2016-2020).}
   \item{median_household_income}{Median household income in past 12 months (2010 dollars, 2006-2010).}
   \item{poverty}{Percent of the population in the county who live in poverty.}
   \item{area}{Total land area, in square miles.}
   \item{density}{Population density, in persons per square mile (2020).}
}
}
\source{
\url{https://www.openintro.org/stat/}
}
\usage{
counties
}
\description{
Data for all the counties in the United States, including variables for demographic,
financial, education, and other characteristics.
}
\details{
Data Source: US Census website.
}
\keyword{datasets}
