\name{hanoverbase-package}
\alias{hanoverbase-package}
\alias{hanoverbase}
\docType{package}
\title{Hanover College Base R package}
\description{
\packageDescription{hanoverbase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hanoverbase}
\packageIndices{hanoverbase}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hanoverbase}

Maintainer: \packageMaintainer{hanoverbase}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
